requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
dnspython==2.2.1
pymongo==3.12.3
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
